requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile /workspaces/cadmus/requirements.in
#
beautifulsoup4==4.11.1
    # via extract-msg
boto3==1.26.35
    # via -r /workspaces/cadmus/requirements.in
botocore==1.29.35
    # via
    #   boto3
    #   s3transfer
cffi==1.15.1
    # via cryptography
chardet==5.1.0
    # via extract-msg
colorclass==2.2.2
    # via oletools
compressed-rtf==1.0.6
    # via extract-msg
cryptography==38.0.4
    # via msoffcrypto-tool
easygui==0.98.3
    # via oletools
ebcdic==1.1.1
    # via extract-msg
extract-msg==0.38.4
    # via -r /workspaces/cadmus/requirements.in
imapclient==2.3.1
    # via extract-msg
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lark-parser==0.12.0
    # via rtfde
lxml==4.9.2
    # via python-docx
msoffcrypto-tool==5.0.0
    # via oletools
olefile==0.46
    # via
    #   extract-msg
    #   msoffcrypto-tool
    #   oletools
oletools==0.60.1
    # via
    #   pcodedmp
    #   rtfde
pcodedmp==1.2.6
    # via oletools
pdfminer==20191125
    # via -r /workspaces/cadmus/requirements.in
pycparser==2.21
    # via cffi
pycryptodome==3.16.0
    # via pdfminer
pyparsing==2.4.7
    # via oletools
python-dateutil==2.8.2
    # via botocore
python-docx==0.8.11
    # via -r /workspaces/cadmus/requirements.in
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
red-black-tree-mod==1.20
    # via extract-msg
rtfde==0.0.2
    # via extract-msg
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   imapclient
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tzdata==2022.7
    # via pytz-deprecation-shim
tzlocal==4.2
    # via extract-msg
urllib3==1.26.13
    # via botocore
